refactor: quoter utilities into utils.tsx
